exclude unnecessary tests to speedup unit tests
Since we are not using many of the plugins and services. We can safely
ignore related tests. Execute './tools/get_ut_exclude_regexp' to get
the regex pattern for exluded unit tests.

Previously it takes about half an hour to finish unit tests on my
desktop. With this patch, it only takes about 5 min.

Closes-bug: OPENSTACK-156