Flating IP rate limit
Add rate_limit to floating IP, then the user could set/update
the bandwidth of a floating IP more convenient.
The valid value of rate_limit is [0, 10000], and the unit
is Mb/Mbit for burst/rate, so the minimum bandwidth will
be 1Mbps.
The cloud user could set the rate_limit in floatingip_create
or update it after creation.

The 0 value of the rate_limit means the unlimited bandwidth.
But only the admin user could set the floating IP rate_limit
value to 0.

Change-Id: I4efe9e49d268dffeb3df4de4ea1780152218633b