Fixes port device_id/device_owner change in failed operation
If a router is a attached to a neutron port and that operation
fails at a late stage in the processing then the 'device_id' and
'device_owner' attributes of the port are still changed. The
failed operation should not change them.

The reason is, as of commit 0947458018725b241603139f4ec6f92e84b2f29b,
the update of those attributes happens outside of the DB transaction.
For failed operation, the values are not reset back to their original
values.

Closes-Bug: #1523859
(cherry picked from commit cb8af9bd50659daca4c1dea4070a011190ef3461)

Conflicts:
	neutron/tests/unit/extensions/test_l3.py
Change-Id: I797df266dafc41843408dc95a6ce9f986db5c21c