When deleting floating IP catch PortNotFound
If we try to delete a VM and to delete the floating IP
associated with the VM at the same time, depending
on the order according to which these requests are processed
Neutron might fail in the deletion of the floating IP,
raising a PortNotFound error. This happens because Neutron
notifies Nova of the network change event and it tries to get
the port to which the FIP is associated. If the port is not there,
Neutron shouldn't raise, it shouldn't send any notification.

Conflicts:
	neutron/notifiers/nova.py
	neutron/tests/unit/notifiers/test_nova.py

Closes-bug: #1586931
Change-Id: Ic72313ad1f787b3cb528e806c843f1fd01eb12f2
(cherry picked from commit 6e275e38575981386508718bf31e6f88b5a0887d)
(cherry picked from commit 9e403b18f03362872571c1b595df9567a7c39e29)