Enable advertising router gateway routes
The neutron DR agent only advertise floating IP routes
as destination(floating IP/32) - next_hop(gateway IP).
Such routes may cause link unreachable in a layer 3
isolated multi-AZ network environment.

This patch will add the router gateway routes to the
bgp speaker routing rules. For instance, a router has
gateway IP 1.1.1.1, and floating IP 2.2.2.2. And this
router is now active in L3 agent, e.g. a master router
for a HA router, which set the external_device_ip to
3.3.3.3 in the br-ex bridge connected to a physical
interface.
So the route rules will be:
+-----------------+--------------+
| destination     | next_hop     |
+-----------------+--------------+
| 2.2.2.2/32      | 1.1.1.1      |
| 1.1.1.1/32      | 3.3.3.3      |(new)
+-----------------+--------------+

And for DVR router, this patch will also add the routes
for the floatingip_agent_gateway in the compute host.
Assuming that in the compute node the fip namespace has
the fg-device with ip 4.4.4.4, and a floating IP bonded
to the VM in this host is 5.5.5.5. The L3 agent in this
compute host set external_device_ip to 6.6.6.6,
then the routes will be:
+-----------------+--------------+
| destination     | next_hop     |
+-----------------+--------------+
| 5.5.5.5/32      | 4.4.4.4      |
| 4.4.4.4/32      | 6.6.6.6      |(new)
+-----------------+--------------+

Change-Id: Ia8d664cc59dcef45314e52e681dba28a4be345a6
(cherry picked from commit 8dc253a2e585ba43d14401307b41d37a0ca6237c)