Add and trigger HA router state change for BGP plugin
When a HA router state change happened, we need to let
the DR agent to know this change as soon as possiable.
So this patch adds a HA_ROUTER resource notify to HA
router plugin and a callback to BGP plugin.
Then when neutron server detected the HA state change,
it will only query and notify the relevant HA router
with the latest gateway routes to the DR agents.

Change-Id: I3906499d98a36927f8d4f3995726c959c00d5a1a